Add a `--resample` feature and flag
As the loudnorm filter might silently upsample the audio if it falls
back to dynamic normalization, a manual resample is required to
keep the rate in the reasonable range. This adds a basic downsampling
filter with a `-r` flag.

TODO: Determine and use the original samplerate as the `aresample` arg.